+++ src/main/java/com/cgd/tinkConnector/Utils/ConversionUtils.java
@@ -3,7 +3,7 @@ package com.cgd.tinkConnector.Utils;
 public class ConversionUtils {
 
 
-    public static float formatAmmount(long amount) {
+    public static float formatAmmount(float amount) {
 
         return (float) (amount / (float) 100);
